requirements.txt
numpy
notebook
pandas
matplotlib
python-mnist